data-ml-pipelines/projects/vendor_ranking/two_towers_v1/requirements.txt
mlflow==2.16.0
tensorflow==2.15.0
tensorflow-recommenders==0.7.3
keras==2.15.0
pyyaml==6.0.1
pydantic==2.7.1
pydantic_settings==2.2.1
python-dotenv==1.0.1
keyrings-google-artifactregistry-auth==1.1.2
google-cloud-storage==2.16.0
cloudpickle==3.0.0
google-cloud-bigquery==3.23.0
pandas==2.2.2
python-geohash==0.8.5
numpy==1.26.4
tqdm==4.66.2
google-cloud==0.34.0
google-cloud-core==2.4.1
db-dtypes==1.2.0
google-cloud-bigquery-storage==2.22.0
pyarrow==10.0.1
fastavro==1.9.4
fsspec==2024.6.1
fastparquet==2024.5.0
gcsfs==2024.6.1
feast==0.40.0
structlog==24.1.0
jinja2==3.1.4
mlserver==1.6.0
mlserver-mlflow==1.6.0
newrelic==9.13.0
requests==2.32.3
evidently==0.4.40